CR: cli: incorrectly specified commands trigger exitCode=1; new "help" command prints usage without the error exit code
